test: added e2e test
