+++ go/test/endtoend/vtgate/queries/dml/sharded_schema.sql
