+++ go/test/endtoend/vtgate/queries/dml/unsharded_schema.sql
