+++ go/test/endtoend/vtgate/queries/dml/vschema.json
